+++ src/mailman/commands/tests/test_cli_inject.py
@@ -62,7 +62,8 @@ class TestInject(unittest.TestCase):
         self.assertEqual(result.exit_code, 2)
         self.assertEqual(
             result.output,
-            'Usage: inject [OPTIONS] LISTSPEC\n\n'
+            'Usage: inject [OPTIONS] LISTSPEC\n'
+            'Try "inject --help" for help.\n\n'
             'Error: No such list: bee.example.com\n')
 
     def test_inject_no_such_queue(self):
@@ -71,7 +72,8 @@ class TestInject(unittest.TestCase):
         self.assertEqual(result.exit_code, 2)
         self.assertEqual(
             result.output,
-            'Usage: inject [OPTIONS] LISTSPEC\n\n'
+            'Usage: inject [OPTIONS] LISTSPEC\n'
+            'Try "inject --help" for help.\n\n'
             'Error: No such queue: bogus\n')
 
     def test_inject_no_filename_option(self):